+++ src/misc/InterfacesAggregated.sol
@@ -126,7 +126,7 @@ interface IZivoeITO is GenericData {
     function claim() external returns (uint256 _zJTT, uint256 _zSTT, uint256 _ZVE);
     function start() external view returns (uint256);
     function end() external view returns (uint256);
-    function stables(uint) external view returns (address);
+    function stables(uint256) external view returns (address);
     function stablecoinWhitelist(address) external view returns (bool);
 }
